Fix VS X64 Compiler Failure Signed-off-by: Liming Gao <liming.gao@intel.com>
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@13113 6f19259b-4bc3-4df7-8a09-765794883524

+++ IntelFrameworkModulePkg/Library/LzmaCustomDecompressLib/F86GuidedSectionExtraction.c
@@ -183,7 +183,7 @@ LzmaArchGuidedSectionExtraction (
   if (!EFI_ERROR (Status)) {
     Status = LzmaUefiDecompressGetInfo (
              Source,
-             SourceSize,
+             (UINT32) SourceSize,
              &OutputBufferSize,
              &ScratchBufferSize
              );
